Bringing in basic files for a game
From my last project called Snake which is chapter 5 of the book, and isn't in fact snake.
Committing main which calls Game which uses the EventManager and the StateManager to organise the state, which inherit BaseSate, which displays in the Window. The keys file is key combos linked to particularly named actions